requirements.txt
llama-index
python-dotenv
html2text
google-generativeai


scikit-learn
